Add conversion functions
Also added several small improvements

+++ backend/PsyTest/PsyTest/Result.cpp
@@ -44,7 +44,7 @@ PsyTest::Result::Result(const Result& obj)
 	if (this->img_size > 0)
 	{
 		this->img = new char[this->img_size];
-		for (int i = 0; i < this->img_size; ++i)
+		for (unsigned long i = 0; i < this->img_size; ++i)
 		{
 			this->img[i] = obj.img[i];
 		}
